refactor(server): remove sync_repo backend, extract GitHub Star endpoint to dedicated route
- Delete routes/sync.py (15 endpoints, 431 lines) — entire sync_repo API surface
- Delete services/multidevice_sync_service.py — 8 classmethods, no longer used
- Delete services/github_sync_scheduler.py — 5-minute auto-sync scheduler, removed
- Add routes/github_star.py — extract /api/github/{starred,star} from former routes/sync.py:243-271, preserves GitHub Star feature decoupled from sync_repo
- routes/__init__.py: drop sync_router, register github_star_router
- routes/dashboard.py: drop sync_repo.get_last_synced_at import + "last synced" display
- routes/settings.py: drop MainConfigResponse.sync_repo field + sync_repo_url write branch
- services/__init__.py: drop MultiDeviceSyncService re-export
- app.py: drop github_sync_scheduler import + startup hook

GitHub Star endpoint paths moved from /api/sync/github/* to /api/github/* — client-side url paths updated in api/client.ts (see prior commit).
