+++ src/frago/server/routes/__init__.py
@@ -9,7 +9,7 @@ This package contains FastAPI routers for different API domains:
 - agent: Agent task execution
 - skills: Claude Code skills
 - settings: Main config, env vars, GitHub integration
-- sync: Multi-device sync via GitHub
+- github_star: Star/unstar the frago repository
 - init: Web-based initialization (dependency check, resource install)
 - viewer: Content preview file serving
 - guide: Tutorial and FAQ content
@@ -23,7 +23,7 @@ from frago.server.routes.agent import router as agent_router
 from frago.server.routes.config import router as config_router
 from frago.server.routes.skills import router as skills_router
 from frago.server.routes.settings import router as settings_router
-from frago.server.routes.sync import router as sync_router
+from frago.server.routes.github_star import router as github_star_router
 from frago.server.routes.init import router as init_router
 from frago.server.routes.viewer import router as viewer_router
 from frago.server.routes.files import router as files_router
@@ -42,7 +42,7 @@ __all__ = [
     "config_router",
     "skills_router",
     "settings_router",
-    "sync_router",
+    "github_star_router",
     "init_router",
     "viewer_router",
     "files_router",
