+++ src/frago/server/services/__init__.py
@@ -10,7 +10,6 @@ from frago.server.services.config_service import ConfigService
 from frago.server.services.env_service import EnvService
 from frago.server.services.github_service import GitHubService
 from frago.server.services.main_config_service import MainConfigService
-from frago.server.services.multidevice_sync_service import MultiDeviceSyncService
 from frago.server.services.recipe_service import RecipeService
 from frago.server.services.skill_service import SkillService
 from frago.server.services.sync_service import SyncService
@@ -27,7 +26,6 @@ __all__ = [
     "EnvService",
     "GitHubService",
     "MainConfigService",
-    "MultiDeviceSyncService",
     "RecipeService",
     "SkillService",
     "SyncService",
